fix listitem
